requirements.txt
# requirements.txt
python-dotenv
huggingface_hub
Pillow